Assemberist improvements
1) There is 3 mode: recording, executing, and waiting of input.
Trigger for entering into first 2 mode is click on button at 6/7 port.
We can't step directly from recording mode to executing mode and vice
versa. So there is not needed check in executing state if we exited from
recording state and reconfigure variables. It can be done at the end
of recording state.

2) Some duplicated code changed on macro.